Move comments to component
